backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
python-dotenv
cors
langchain
langchain-google-genai
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pydantic-settings
python-jose[cryptography]
